- upload 2nd milestone

+++ sectors/module/api2file.py
@@ -76,10 +76,16 @@ class Bridge:
 
     def write_file(self, data):
         self.add_cache(f'FILE:Update - {data}')
+        bridge = TBLBridge.objects.get(id=self.bridge_info['id'])
+        if bridge.is_status == 1 or bridge.user.balance <= 0:
+            bridge.is_status = 1
+            bridge.save()
+            self.add_cache(f'FILE:Update - Ignored! - Out of Funds!')
+            return
+
         self.file.truncate()
         self.file.write(data)
 
-        bridge = TBLBridge.objects.get(id=self.bridge_info['id'])
         bridge.api_calls += 1
         bridge.save()
